piano_interface/requirements.txt
numpy==1.25.2
opencv_python==4.8.0.74
PyQt5==5.15.9
PyQt5_sip==12.12.2
tensorflow_addons==0.19.0
tensorflow_intel==2.10.0
